+++ pnet-data-api-java/src/main/java/pnet/data/api/company/CompanyDataAutoComplete.java
@@ -11,6 +11,7 @@ import pnet.data.api.util.RestrictBrand;
 import pnet.data.api.util.RestrictContractState;
 import pnet.data.api.util.RestrictContractType;
 import pnet.data.api.util.RestrictCountryCode;
+import pnet.data.api.util.RestrictDatedBackUntil;
 import pnet.data.api.util.RestrictLocation;
 import pnet.data.api.util.RestrictTenant;
 import pnet.data.api.util.RestrictType;
@@ -25,7 +26,7 @@ public class CompanyDataAutoComplete extends AbstractAutoComplete<CompanyAutoCom
     RestrictCountryCode<CompanyDataAutoComplete>, RestrictType<CompanyDataAutoComplete>,
     RestrictContractType<CompanyDataAutoComplete>, RestrictContractState<CompanyDataAutoComplete>,
     RestrictLocation<CompanyDataAutoComplete>, IncludeInactive<CompanyDataAutoComplete>,
-    CompanyMergable<CompanyDataAutoComplete>
+    RestrictDatedBackUntil<CompanyDataAutoComplete>, CompanyMergable<CompanyDataAutoComplete>
 {
 
     public CompanyDataAutoComplete(AutoCompleteFunction<CompanyAutoCompleteDTO> autoCompleteFunction,