+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonDataAutoComplete.java
@@ -13,6 +13,7 @@ import pnet.data.api.util.RestrictCompany;
 import pnet.data.api.util.RestrictCompanyId;
 import pnet.data.api.util.RestrictCompanyNumber;
 import pnet.data.api.util.RestrictCredentialsAvailable;
+import pnet.data.api.util.RestrictDatedBackUntil;
 import pnet.data.api.util.RestrictFunction;
 import pnet.data.api.util.RestrictTenant;
 
@@ -26,7 +27,8 @@ public class PersonDataAutoComplete extends AbstractAutoComplete<PersonAutoCompl
     RestrictCompanyNumber<PersonDataAutoComplete>, RestrictCompany<PersonDataAutoComplete>,
     RestrictFunction<PersonDataAutoComplete>, RestrictActivity<PersonDataAutoComplete>,
     RestrictCredentialsAvailable<PersonDataAutoComplete>, RestrictApproved<PersonDataAutoComplete>,
-    IncludeInactive<PersonDataAutoComplete>, CompanyMergable<PersonDataAutoComplete>
+    IncludeInactive<PersonDataAutoComplete>, RestrictDatedBackUntil<PersonDataAutoComplete>,
+    CompanyMergable<PersonDataAutoComplete>
 {
 
     public PersonDataAutoComplete(AutoCompleteFunction<PersonAutoCompleteDTO> autoCompleteFunction,